Better handle the memory by allocating ogrid instead of mgrid
In the case of large stimuli (512**3 for instance), we waste a lot of
memory by allocating many big matrices. As a start, we give the
possibility of using ogrid instead of mgrid. This won't work if you want
to use Mayavi visualisations (they need non-sparse input for a
scalar_field) but will work to simply generate the stimuli (for instance
as a mat file). This behaviour is by default and should just not change
anything else.