Using a more compact formulation for the von-mises distribution
The orientation is between O and \pi and the Fourier spectrum has a
double peak, symetrical centrally around the origin.

We therefore use here the von-Mises distribution on the [O, pi) interval
by doubling the angle. This gives very similar results (see
test_orienttation.py) but with a more physical writing than the one on
the JNP paper.